pyproject.toml
[tool.poetry]
name = "qzandronum_site"
version = "0.42.1-b-1"
description = "Django website for Qzandronum project"
authors = ["Gin Fuyou <devel@doratoa.net>"]

[tool.poetry.dependencies]
python = "^3.10.4"
pre-commit = "~2"
django-classy-settings = "^2.1"
Django = "3.2.*"
django-debug-toolbar = "~2"
celestia = { git = "https://gitlab.com/GinFuyou/django-celestia.git" }

pip = ">=20.2.4"
bleach = "^3.2"
django-sniplates = "^0.7.0"
uWSGI = "^2.0.19"
psycopg2 = "^2.8.6"
django-chunked-upload = "^2.0.0"
httpagentparser = "^1.9.5"


[tool.poetry.dev-dependencies]
#celestia = {path = "../celestia", develop = true}
isort = {extras = ["pyproject"], version = "^5.6.4"}
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
